+++ src/auth/containers/forgotPassword/ForgotPassword.tsx
@@ -14,6 +14,8 @@ import { SetAuthenticated } from "../../../app/core/redux/app/actions";
 
 import ForgotPasswordForm from "../../components/forgotPassword/ForgotPasswordForm";
 
+import { getCurrentSessionTokens } from "../../core/services/session";
+
 const mainImage = require("../../../shared/assets/images/main-image.jpg");
 
 const ForgotPassword = (props: any) => {
@@ -21,11 +23,13 @@ const ForgotPassword = (props: any) => {
 
   const { isAuthenticated } = props.state.auth;
 
+  const { accessToken } = getCurrentSessionTokens();
+
   useEffect(() => {
     if (isAuthenticated) {
       props.history.push("/dashboard");
     }
-    if (localStorage.getItem("accessToken")) {
+    if (accessToken) {
       setAuthenticated();
     }
   });
