+++ src/auth/containers/login/Login.tsx
@@ -16,6 +16,8 @@ import LoginForm from "../../components/login/LoginForm";
 
 import { SetAuthenticated } from "../../../app/core/redux/app/actions";
 
+import { getCurrentSessionTokens } from "../../core/services/session";
+
 const mainImage = require("../../../shared/assets/images/main-image.jpg");
 
 const Login = (props: any) => {
@@ -23,11 +25,13 @@ const Login = (props: any) => {
 
   const { isAuthenticated } = props.state.auth;
 
+  const { accessToken } = getCurrentSessionTokens();
+
   useEffect(() => {
     if (isAuthenticated) {
       props.history.push("/dashboard");
     }
-    if (localStorage.getItem("accessToken")) {
+    if (accessToken) {
       setAuthenticated();
     }
   });
