+++ src/auth/core/services/session.ts
@@ -1,0 +1,11 @@
+export const getCurrentSessionTokens = () => {
+  const accessToken = localStorage.getItem("accessToken")
+    ? localStorage.getItem("accessToken")
+    : sessionStorage.getItem("accessToken");
+
+  const refreshToken = localStorage.getItem("refreshToken")
+    ? localStorage.getItem("refreshToken")
+    : sessionStorage.getItem("refreshToken");
+
+  return { accessToken, refreshToken };
+};
